Refactored cargo utilities
