+++ bin/checks.ts
@@ -4,11 +4,12 @@ import fs from 'fs'
 import util from 'util'
 import yargs from 'yargs'
 
-import { isCargoBuildSuccessful, tryCargoRun } from '../src/checks/cargo/cargo'
+import { isCargoBuildSuccessful } from '../src/checks/cargo/cargo'
 import { cargoCheckCheck } from '../src/checks/cargo/cargo-check'
 import { cargoClippyCheck } from '../src/checks/cargo/cargo-clippy'
 import { cargoFmtCheck } from '../src/checks/cargo/cargo-fmt'
 import { cargoTestCheck } from '../src/checks/cargo/cargo-test'
+import { tryCargoRun } from '../src/checks/cargo/run-cargo'
 import { runCargoCheck } from '../src/checks/cargo/run-cargo-check'
 import { runCargoClippy } from '../src/checks/cargo/run-cargo-clippy'
 import { runCargoFmt } from '../src/checks/cargo/run-cargo-fmt'
